Clean up, move field to another module for global access.
